Socioeconomic variables database
Three files.

- First one ("Index") contains diferent index calculated by Madrid City Council to study vulnerability aspects in each district and neighbourhood/suburb. 

- Second one ("Indicators") contains the variables used to calculate the index mentioned before. As you will see, some variables have neighbourhoods/suburb data while others use data from a district to all their neighbourhoods.

- Third one, "Environmental label", contains the percentage of vehicles for each neighbourhood with environmental label and, consequently, allowed access to the LEZ.

I don't know if it is interesting to use all the variables in our models or if it is better to remove some of them (for example, average time arrival for ambulances don't seem to be much important to define vehicle environmental label). I'll try to also do a test with only economy and employment variables.